Splitting up the tests
